feat Addition of endpoints and methods for Administrators
New endpoints have been added to the `Endpoint.cs` file to manage administrators, including creation, retrieval of all, and retrieval by ID.

Updated the `IServiceAdministrator` interface with new methods to support the operations of creating, retrieving all, retrieval by ID, and validation of administrator data.

Implemented the `Create`, `FindById`, `GetAll`, and `Validate` methods in the `ServiceAdministrator` class to allow creation, search, paged listing, and validation of administrators.

Created the `AdministratorDTO` class to define the structure of administrator data in requests.

Created the `Profile` enum to define the possible profiles for an administrator (`Adm` and `Editor`).

Implementation of `CreateAdministratorEndPoint`, `GetAdministratorByIdEndPoint` and `GetAdministratorEndPoint` endpoints for creating, getting by ID and paged getting of administrators, respectively.